add generic table component
