+++ pnpm-workspace.yaml
@@ -19,8 +19,8 @@ catalog:
   '@nuxt/module-builder': ^0.8.4
   '@nuxt/schema': ^3.15.0
   '@nuxt/test-utils': ^3.15.4
-  '@nuxt/ui-pro': 3.0.0-alpha.12
-  '@nuxt/ui': 3.0.0-alpha.12
+  '@nuxt/ui-pro': https://pkg.pr.new/@nuxt/ui-pro@37c0e89
+  '@nuxt/ui': https://pkg.pr.new/@nuxt/ui@4aa3179
   '@nuxtjs/supabase': ^1.4.5
   '@release-it/conventional-changelog': ^10.0.0
   '@tailwindcss/typography': ^0.5.15
